Created the concept of LIMITERs
Since the class methods of receives, requires and the yet to be
implemented returns all behave the same it felt knida redundant to
define and write tests everytime one was added.
All the methods are now metaprogrammed instead and are testest in a loop
as well.

Limiters seemed the most fitting name at the time being, but theres
always the posibility of it to change.