Created two module to manage method missings
The idea was to divide method missing in 2 parts since the Result should
not be able to modify anything.

The InstanceVariableReadable module helps respond to undefined READ
methods when an instance variable is already defined. This will also
create the methods in the class so method_missing is just called once.

The InstanceVariableWriteable module helps respond to assignement
methods, basically anything that looks like one lol.