Refactored Limitable, added skip_validations
So after having the realization that I should probably unbind the
validation from the initialization of the command (will make hooks
viable) Limitable got refactored once again, this time everything ends
up in the same scope and for the first time, it's actually readable.

skip_validations option is now added as well, so now you can skip the
validations when calling the command if for whathever you so desire.

Happiness and joy.