Refactored InstanceVariable Missing modules
It felt just very awkward to have not one, but two modules names
InstanceVariableSomething, so just renamed them to IvarSomething instead
which fortunately is still easy to remember and shorter.

Also slightly improved performance on some of the methods to use simple
comparisons instead of regex. I know I really like using them but
definitely need to consider performance of the others.