Added Shareable feature
This allows commands to be shareable down the line.
In theory this will allow you to add functionality (hooks, limits,
status) to modules or with classic inheritance.

Yes, class instance variables are being copied over to new classes. This
would actually not be an ok approach if it were not because in theory
Command classes are a one method object. Polluting the classes too much
is not too pretty.

The other downside that comes with this approach is that the
hooks/limits/suceeds status that are added will only be added in load
time. Once you've inherited if you dynamically add something to the
parent class or module it will be 'stuck' there. If such is your case,
you probably should use ActiveSupport::Concern's approach instead.