Changed how Commander initialize and result work
Previously a commander would receive a bunch of arguments and save them
to later initialize the first command to be executed.
It also would use after and aborted hooks to clean up the results.

The main issue here is that it added a lot of inconsistency for the
uses. When is it ok to use command? When is it ok to no longer use it?
This is specially true with hooks since some would have command, and
some wouldn't.

In an effort to standardize, now a Commander will always have a command
after being initialized. At this point it is an empty "Dummy" command
to which the user can add variables to pass to the next one.

Instead of preparing the result by hooks, the cleanup now happens after
running the hooks, so .call was tweaked a bit at class level.