Commander now skips :receives validations of commands
This change was made because since the commander returns the complete
set of results to the next command more often than not it would conflict
with the receives option.

Since receives is a "typo correctness" kind of validation rather than a
"I need this to work" one I thought this would be the best approach.

The other available approach is of course, cleaning after every command
and reenabling the receives validation, but I feel this would be a bit
overly complex for the user. We'll see how the feedback goes.