:bug: Fix an issue when inheriting the command
Just found a fresh one. I would be surprised this was not found before
if someone actually used the gem. *sadface

Basically when using inheritance the limiters/succeeds/hooks were
pointing to the same object id so when you updated one it propagated to
the others. I'm a bit interested on why I was not able to reproduce this
behaviour with mixins but a fix was applied just in case nontheless.